Don't return nulls in offsetsForTimes
The Java Kafka API will put null values in the map returned by
offsetsForTimes, for partitions that don't have any messages in them
yet. This is the case, even the documentation only specifies
null-behavior for "old" message formats.

This commit filters out the null values, since using null in a scala Map
is very un-idiomatic, and would be surprise to most users.

There should be no loss of information, since the caller has the
complete list of partitions in the method argument.