adds failure in case of resource leakage
It was possible for transactions to leak, which is particularly disruptive in case transactions are often open. This is because, while normally producing outside a transaction throws an exception, producing with an old transaction while a new one is open doesn't result in any error.
This commit should enforce transaction exhaustion once it's either committed or aborted, signaling to the user whenever a transaction has been leaked.