Add support for transactions through a TransactionalProducer (#356)
* transactional producer second draft

* added tests

* fmt

* working with offsets

* hide TransactionalProducer constructor

* hide TransactionalProducer constructor

* abstract Transaction

This might be useful since TransactionalProducer is a module and depends
on Transaction. This means that a user who wants to implement a test version
needs access to subclasses of Transaction.

* removed deprecated method

https://github.com/apache/kafka/blob/f650a14d56c0cc33263c29d8d242760406943c5b/clients/src/main/java/org/apache/kafka/clients/producer/Producer.java#L53-L55
favor other non-deprecated method

* fmt

* adds failure in case of resource leakage

It was possible for transactions to leak, which is particularly disruptive in case transactions are often open. This is because, while normally producing outside a transaction throws an exception, producing with an old transaction while a new one is open doesn't result in any error.
This commit should enforce transaction exhaustion once it's either committed or aborted, signaling to the user whenever a transaction has been leaked.

* fmt

* mark as closed before anything else

* fix scala 2.12.14 compilation on ci

* optimize in case of no offset to commit