Wait for stream end in rebalance listener and on unsubscribe
This PR adds the setting `endRevokedStreamsBeforeRebalance`. When set to `true` (the default) a revoked partition is kept until the stream is ended. This gives the streams an opportunity to commit all offsets before another client gets the partition assigned. This prevents duplicate processing.

When `endRevokedStreamsBeforeRebalance` is set to `false` the previous behavior is active. This lets the old stream finish its work in the background while the new stream is already starting. The old behavior causes duplicate processing but can be a bit more performant.

In order to handle commits while waiting for the streams to end, commits are added to a separate queue. To make sure the main poll loop knows about these commits, we also place a `CommitAvailable` value on the command queue.

The rebalance listener ZIO effects (and other callbacks) are executed on the calling thread. This is required because the underlying Java client requires single threaded access. It rejects multiple concurrent invocation unless they are from the same thread, that is, the thread that executes the poll. (Remember, the rebalance listener and other callbacks are invoked by the Java client during a call to poll or commit.)

While waiting for the streams to end we need to continuously poll the broker. We do this by calling `commitAsync`. By inspecting the Java client I found that when the passed Map of offsets is empty, the client will still poll the broker and call callbacks. This is exactly what we need.

The rebalance listener no longer manipulates the state. The rebalance event only collects changes. The main poll loop then use those changes to construct the next state. When the rebalance listener was not invoked, we skip some tasks for a small performance improvement.

Also:
 - End all streams when only onLost is invoked.
 - Disable zio-intellij's inspection `SimplifyWhenInspection` in Runloop because its suggestion is not equivalent (performance-wise).
 - Configure a low `max.poll.records` (100 instead of 1000) for tests. There was a concurrency bug that only failed a test the lower setting. The benchmarks continue to use 1000 for `max.poll.records`.
 - Prevent deadlock by doing unsubscribe with a single consumer access.
 - Await stream end on unsubscribe and shutdown as well.
 - Removed method `Runloop.awaitShutdown`, it is not used, not accessible and its semantics are questionable.