Wait for stream end in rebalance listener
This PR adds the setting `endRevokedStreamsBeforeRebalance`. When set to true rebalances are postponed until the streams that handle the revoked partitions have ended. This prevents duplicate processing.

In order to handle commits while waiting for the streams to end, each commit is added to 2 queues. The command queue (as before) for handling commits after calling `poll`, and the commit queue for handling inside the rebalance listener. Each commit can be executed only once.

While waiting for the streams to end we need to continuously poll the broker. We do this by calling `commitAsync`. By inspecting the Java consumer code I found that when the passed Map of offsets is empty, the Java consumer will still poll the broker which is exactly what we need.

The main runloop no longer needs to get a detailed result from the rebalance listener. Just knowing that the listener was invoked is enough to go detect the changes. When it was not invoked, we skip those checks for a small performance improvement.

TODO: an attempt is made to make the rebalance listener ZIO effect be executed on the calling thread. Unfortunately this does't work yet. Getting this to work is a requirement for this PR.