Replace pending requests by single boolean
Since streams now wait for data by listening to their own queue, we no longer need to keep track of pending requests. Instead:
- streams initiate a poll when they need data,
- runloop initiates another poll when not all streams got data.

This also removed field `tpRequested` from `DiagnosticEvent.Poll` since it simply the union of fields `tpWithData` and `tpWithoutData`.

Runloop's `State` is moved into the `Runloop` object to hide it for the rest of the package.