Add processing timeout at the stream level (#1009)
Kafka's process to detect halted consumers is crude. Because it is based on polls, it works per consumer. Here we make it more fine-grained by doing slow consumer detection per stream.

For the Kafka client 'halting' means poll is not called for longer than `max.poll.interval.ms`. Translated to zio-kafka 'halting' means that a stream does not consume any records for `max.poll.interval.ms` or longer, while records are available.

The halted stream is interrupted so that it can fail properly. In addition, we shut down the entire consumer. This will help the broker to detect slow consumers. Ideally we would only stop the subscription that the stream belongs to. Unfortunately, we do not maintain a topic belongs-to subscription relation. Therefore, stopping only the affected subscription is left for a future change.

Fixes #902.