Keep track of last pulled offset (#1086)
For #830 we need to keep track of which records were pulled from the stream so that we can wait for them to be committed. This change prepares for that.

The idea is that the rebalance listener first ends the streams for revoked partitions, then it will get the last consumed offset using the now public `completedPromise`. With that info, it can await these commits to complete.

The advantage of this approach is that it is very precise; we only await offset commits for records that we know were pulled from the stream. Alternative implementation would track the given offsets inside Runloop where we do not have the exact knowledge of what was pulled or not.