Track latest completed commit offset per partition (#1097)
By tracking these offsets we can skip awaiting already completed commits from the rebalance listener in #830.

To prevent unbounded memory usage, after a rebalance we remove the committed offset for partitions that are no longer assigned to this consumer.

Note that a commit might complete just after a partition was revoked. This is not a big issue; the offset will still be removed in the next rebalance. When the `rebalanceSafeCommits` feature is available and enabled (see #830) commits will complete in the rebalance listener and this cannot happen anymore.

The offsets map is wrapped in a case class for 2 reasons:
* It provides a very nice place to put the updating methods.
* Having updating methods makes the code that uses `CommitOffsets` very concise.