Protect against slow user diagnostics
Diagnostics is a feature of zio-kafka that allows users to listen to key events. Since zio-kafka calls out to the user's implementation of the `Diagnostics` trait, there are no guarantees on how well it behaves.

This is even more important inside the rebalance listener where we (soon, with #1098) run on the same-thread-runtime and can not afford to be switched to another thread by ZIO operations that are normally safe to use.

To protect against these issues:
 - the rebalance events are replace by a single event which is emitted from outside the rebalance listener,
 - all invocations of the diagnostics trait are forked (unless they are run from a finalizer).