Pause a partition when its stream is ended
During a graceful shutdown the partition stream is ended. The runloop however, is still happily fetching records for that stream. These are then put in the stream's queue even though they will never be read. This is taking away network and CPU at a moment where the application should focus on doing a clean shutdown quickly. It could even cause an OOM for application that are tuned for the case where processing happens almost immediately.

With this change, when a stream is ended:
- the partition is paused,
- already fetched new records are not added to the stream's queue.

Also: less logging at start of a poll.