Allow stream to be interrupted when there is no traffic (#1251)
When a partition is lost while there is no traffic,
`PartitionStreamControl` is blocked waiting for data. We fix this by
racing with the `interruptPromise`. (Thanks @josdirksen for the
analysis! See #1250.)

Note: this situation can only occur with lost partitions. Timeouts (the
other reason for interrupts) do not occur when there is no traffic.
Currently, we have no way to test lost partitions. Therefore, there are
no added tests.

This PR does _not_ change how lost partitions are handled. That is, the
stream for the partition that is lost is interrupted, the other streams
are closed gracefully, the consumer aborts with an error.