Aborts builds when a newer build is ready (#1255)
Right now, we run all the test suites, a full benchmark, and a
memory/CPU profiler on _every commit_. These builds continue, even when
a newer commit is available. This PR reduces this considerably by
cancelling builds when a newer commit was created. Builds on the default
branch ('master') are never cancelled.

The `ci` workflow is generated and therefore not changed in the same
way.