Do not start streams for immediately revoked/lost partitions
As discovered by @ytalashko (see https://github.com/zio/zio-kafka/pull/1294), it is possible that a partition is assigned and immediately revoked in the same poll. No stream should be started for these partitions.

Unlike #1294, this PR does _not_ hide the assigned+revoked partitions from the diagnostic events and rebalance metrics. In addition, this PR also supports the unlikely event of a assigned+lost partition.