Revert back to 2.8.0 sendFromQueue implementation in producer (#1304)
During running services with the new version of library 2.8.1, I noticed
huge increase in messages production time to kafka.
Some quick rough tests shoving me around 40x-100x times increase in
amount of time taken to `produceChunk` on even 1-10 records chunks
comparing to version 2.8.0 (or also to version 2.8.1 with the changes in
this MR). Please, note, it is not a proper benchmarks. Also, to note, Im
using Scala version 3.3.1.

This MR just reverts two MRs updating `ProducersendFromQueue`
implementation:
 - https://github.com/zio/zio-kafka/pull/1272
 - https://github.com/zio/zio-kafka/pull/1285
 
Not sure if it's possible to revert two MRs at a time (with a single one
for revert), so created this one.
 
I haven't researched yet which exact change/changes are causing such
performance degradation.
 I would suggest the next steps:
  - confirm the problem exists
- reverting to the previous implementation (the one from 2.8.0/this MR)
  - release fixed version (to allow users have a nicely working version)
- investigate & fix problem from the
https://github.com/zio/zio-kafka/pull/1272 and/or
https://github.com/zio/zio-kafka/pull/1285

It is only suggestions on the approach, feel free to ignore them.
Also, feel free to modify/ignore this MR and treat it as an issue.