More precise batch publish errors (#1321)
This increases the precision of the result of producer method
`produceChunkAsyncWithFailures` for cases where something went wrong.
There are two changes compared to before:

1. Each entry in the result of`produceChunkAsyncWithFailures` now
accurately corresponds to each record in the input chunk. Previously, if
sending fails directly (*) on any of the given records, the error would
be used for _all_ records in the batch, ignoring the send-outcome of the
other records. An advantage of this change is that if sending some
records failed, but some other records were actually sent, you can now
correctly see all of that in the method's response.
2. In addition, if sending fails directly for a record (*), we no longer
attempt to send subsequent records from the input. The result contains a
`PublishOmittedException` for each record that is not sent. When
implementing retries, this change makes it easier to publish records in
the original order.

In addition, we introduce unit-level tests for the producer.

(*) By 'sending' we mean offering a record to the underlying java Kafka
producer. Sending can fail directly (when we call the method), or later
on (from a callback).