Run profiler only on release and on demand
Since we are not using the cpu/heap profiles that much, it is an environmental offence to run the profiler on every commit.

With this change we only run the profiler on releases. In addition, it is now easy to run the profiler on demand from any branch, tag or commit.