Improve tests by removing sleeps (#1401)
This change improves the reliability of tests in
`ConsumerSpec.scala` by replacing the use of `ZIO.sleep` with
synchronization primitives such as `Promise` and `Ref`. The previous
implementation relied on arbitrary sleep durations to wait for certain
conditions, which could lead to non-deterministic behavior and flakiness
in the tests.

Follow up of #1389.