[Internal] Micro-optimisation: Reuse previously calculated `java.langSystem.nanoTime` (#1407)
Hey :)

I'm not contributing much anymore. Sorry about that. I lost the
motivation, TBH, but I really enjoy following what you continue to do 👏

I was passing by and was reading some of the new code when I saw this.
LMKWYT :)

With this change, we lose some nanoseconds of precision, so it might not
be the expected behaviour. 🤔