Extract Committer and RebalanceCoordinator classes from Runloop + unit tests (#1375)
The code is mostly just moved to a different place, the logic was left
mostly intact. The 'interface' between the components has been decoupled
more, i.e. the rebalance listener no longer access the full Runloop's
State and the pending commits are stored internally in the Committer.

Care has been taken to make the Committer usable during rebalancing as
well, with the proper access to the Consumer for example. The part that
waits the end of the streams and their commits has been changed to use
the Committer.

Includes unit tests for the two new components.

---------

Co-authored-by: Erik van Oosten <e.vanoosten@grons.nl>