Simplify transactional producing with rebalanceSafeCommits
This simplifies the setup for transactional producing and committing, no longer requiring a custom rebalance listener or the use of `partitionedAssignmentStream`. See `ConsumerSpec` for an example.

Note that the transaction ID is no longer used for fencing since https://cwiki.apache.org/confluence/display/KAFKA/KIP-447%3A+Producer+scalability+for+exactly+once+semantics, the mechanism relies on the consumer group metadata instead.

NOTE This change is not binary compatible and is therefore, for zio-kafka 3.0.0