Preparing for producer diagnostics
It has been a long wish to have producer diagnostics. Although these are not really hard to add, the consumer diagnostics are a bit in the way; these already claim some names and make code sharing awkward. Now that we are moving towards zio-kafka 3.0 we can fix these at the cost of some backward incompatibilities.

This PR has the same changes as #1507 (which will be applied to the series/2.x branch) but then with all deprecated code removed.