Explore event sourced rebalance listener
In this experiment we remove even more logic from the rebalance listener; it now reports which callbacks happened and no longer collapses that information. This gives the runloop a full picture of what happened during the rebalance.

For now, the logic stays exactly the same, it just moved to the runloop. However, the idea behind this change is that it will make it easier to support the cooperative partition assigner strategy.