Preparing for producer diagnostics
It has been a long wish to have producer diagnostics. Although these are not really hard to add, the consumer diagnostics are in the way; these already claim some names and make code sharing awkward. Now that we are moving towards zio-kafka 3.0 we can fix this at the cost of
some backward incompatibilities.

This PR prepares for the introduction of producer diagnostics by moving the sharable diagnostics code into a new higher level `diagnostics` package. In addition, zio-kafka internal code is moved to the `diagnostics.internal` package.

This also removes the trait `DiagnosticEvent.Finalization` from the type hierarchy because it doesn't serve any user purpose (it was introduced to make it easier to test the shutdown sequence of the consumer).

All old APIs are still available but are deprecated.

This change is source compatible, but _not_ binary compatible.