Preparing for producer diagnostics (3.x) (#1452)
It has been a long wish to have producer diagnostics. Although these are
not really hard to add, the consumer diagnostics are a bit in the way;
these already claim some names and make code sharing awkward. Now that
we are moving towards zio-kafka 3.0 we can fix these at the cost of some
backward incompatibilities.

This commit has the same changes as commit
1779b27bb06e2ffd5693b6f9674b5616e5f63b32 in the series/2.x branch, but
then with all deprecated code removed.