Provide diagnostics via `ConsumerSettings` (#1515)
Since most people do not use consumer diagnostics, it is a bit annoying
that everybody has to provide a (dummy) diagnostics layer when the
zio-kafka `Consumer` is constructed via a layer.

This is solved by moving the configuration of diagnostics to
`ConsumerSettings`.

This approach also makes it consistent with the to-be-introduced
`ProducerDiagnostics`.