Add producer diagnostics (#1518)
Producer diagnostics allows the user to keep track of what happens in
the producer. It gives information on where records are produced to, and
the size of each record.

Benchmarks show no overhead when these new producer diagnostics are
disabled. Enabling the new producer diagnostics should increase heap
usage somewhat, but we did not see any effects in the benchmark.

Also: improve scaladoc of `ProducerSettings`