Graceful shutdown of a stream for a single subscription (#1201)
Implements functionality for gracefully stopping a stream for a single
subscription: stop fetching records for the assigned topic-partitions
but keep being subscribed so that offsets can still be committed.
Intended to replace `stopConsumption`, which did not support
multiple-subscription use cases.

A new command `EndStreamsBySubscription` is introduced, which calls the
`end` method on the `PartitionStreamControl` of streams matching a
subscription. In the method `Consumer#runWithGracefulShutdown` we then
wait for the user's stream to complete, before removing the
subscription.

This is experimental functionality, intended to replace
`stopConsumption` at some point. Methods with this new functionality are
offered besides existing methods to maintain compatibility.

~All the fiber and scope trickery proved to be very hard to get right
(the lifetime of this PR is a testimony to that), and there may still be
subtle issues here.~ This is now traced back to issue
https://github.com/zio/zio/issues/9288

Implements some of #941.

---------

Co-authored-by: Erik van Oosten <e.vanoosten@grons.nl>
Co-authored-by: zio-scala-steward[bot] <145262613+zio-scala-steward[bot]@users.noreply.github.com>